fix: ajuste na estilização e criação de mock
