+++ src/components/ui/highlight-card/mocks/highlight-card.mock.ts
@@ -1,0 +1,33 @@
+import { IHighlightCardItem } from '../highlight-card.type';
+
+const items: IHighlightCardItem[] = [
+  {
+    type: 'message',
+    title: 'Que bom ver vocês por aqui!',
+    content:
+      'Estamos animados para compartilhar as melhores ofertas, produtos exclusivos e momentos especiais com vocês. Fiquem ligados e aproveitem ao máximo!',
+    highlight: true,
+  },
+  {
+    type: 'product',
+    id: 9480063,
+    price: 799.0,
+    priceBase: 1290,
+    image: {
+      src: 'https://optimizer.dchomolog.dooca.store/103779/products/013886p1a4.jpeg?v=1695153157',
+      alt: null,
+    },
+    name: 'Tênis Nike Air Force',
+    specialPrice: 395.505,
+    highlight: true,
+  },
+  {
+    type: 'message',
+    title: 'Que bom ver vocês por aqui!',
+    content:
+      'Estamos animados para compartilhar as melhores ofertas, produtos exclusivos e momentos especiais com vocês. Fiquem ligados e aproveitem ao máximo!',
+    highlight: false,
+  },
+];
+
+export { items };
