+++ src/index.html
@@ -26,7 +26,8 @@
     >
       <img src="https://picsum.photos/720/340" alt="Banner Image" />
     </custom-card> -->
-    <live-video-player video-id="h5iOu3zAj-Q" auto-play></live-video-player>
+    <!-- <live-video-player video-id="h5iOu3zAj-Q" auto-play></live-video-player> -->
+    <highlight-card></highlight-card>
   </body>
   <!-- <script>
     const btn = document.getElementById('btn-go');
